refactor : line생성에 대한 별도 class 분리 및 내부 로직 수정
